+++ NEWS.md
@@ -1,10 +1,12 @@
 ## propr 3.0.1.9000
 ---------------------
-* Implement "weighted theta" calculation
+* Implement backend for "weighted theta" calculation
     * Added `wtmRcpp` for weighted mean calculation
     * Added `wtvRcpp` for weighted variance calculation
     * Added `lrm` function with optional weighted calculations
     * Added `lrv` function with optional weighted calculations
+* Implement "weighted theta" calculation
+    * Added `calculateThetaW_old` for unit testing
 
 ## propr 3.0.1
 ---------------------